# specs/012-video-transcription/contracts/seminar-recording-api.openapi.yaml
openapi: 3.1.0
info:
  title: Aither-Hemera Seminar Recording Integration
  version: 1.0.0
  description: >-
    Service contracts for validated booking lookup, durable transcription workflow
    state, and authorized short-lived media access. Hemera owns workflow records
    and booking authorization; Aither owns AssemblyAI, MUX, and Blob credentials.
servers:
  - url: https://aither.example.invalid
    description: Aither service (placeholder)
  - url: https://hemera.example.invalid
    description: Hemera service (placeholder)
...
paths:
...
  /api/service/seminar-document-access:
    post:
      tags: [Aither Service API]
      summary: Mint short-lived media access after Hemera booking authorization
      description: >-
        Called only by Hemera server-side after checking the authenticated user's
        ownership of bookingId. This endpoint is not a substitute for Hemera's
        user authorization check.
      security:
        - AitherServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, recordingId]
              properties:
                bookingId:
                  type: string
                recordingId:
                  type: string
      responses:
        '200':
          description: Temporary bearer URLs for a ready workflow
          content:
            application/json:
              schema:
                type: object
                required: [muxPlaybackUrl, transcriptUrl, muxExpiresAt, transcriptExpiresAt]
                properties:
                  muxPlaybackUrl:
                    type: string
                    format: uri
                    description: Fresh signed HLS URL; expiry exceeds the full video duration. This request-time URL is distinct from the stable muxPlaybackUrl stored in Hemera and must not be persisted or logged.
                  transcriptUrl:
                    type: string
                    format: uri
                    description: Private Blob GET URL expiring after exactly 300 seconds.
                  muxExpiresAt:
                    type: string
                    format: date-time
                  transcriptExpiresAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workflow is not ready for participant access
  /api/recording/transcription/{recordingId}/review:
    put:
      tags: [Aither Service API]
      summary: Approve a corrected speaker-role mapping
      description: >-
        Requires an authenticated Aither administrator. The workflow remains
        unavailable to Hemera participants until this correction is validated
        and the publish stages complete.
      security:
        - ClerkSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RecordingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, participantSpeakerId, leaderSpeakerId, approve]
              properties:
                bookingId:
                  type: string
                participantSpeakerId:
                  type: string
                leaderSpeakerId:
                  type: string
                approve:
                  type: boolean
                  const: true
      responses:
        '200':
          description: Mapping accepted; workflow resumed for publication
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not authorized to review transcription
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workflow is not awaiting speaker review
        '422':
          description: Mapping is invalid or maps both roles to the same speaker
...
  /api/assemblyai/webhook:
    post:
      tags: [AssemblyAI Callback]
      summary: Receive an authenticated AssemblyAI transcription completion callback
      security:
        - AssemblyAIWebhookSecret: []
      parameters:
        - name: recordingId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transcript_id, status]
              properties:
                transcript_id:
                  type: string
                status:
                  type: string
                  enum: [completed, error]
      responses:
        '202':
          description: Callback accepted for idempotent canonical-workflow processing; duplicate or stale delivery does not overwrite an approved mapping or completed stage.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Hemera unavailable; callback should be retried
components:
  securitySchemes:
    HemeraApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    AitherServiceKey:
      type: apiKey
      in: header
      name: X-Aither-Service-Key
    AssemblyAIWebhookSecret:
      type: apiKey
      in: header
      name: X-AssemblyAI-Webhook-Secret
    ClerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    RecordingId:
      name: recordingId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid service credential
    NotFound:
      description: Booking or recording workflow not found